pyproject.toml
[tool.poetry]
name = "telebot-components"
version = "0.0.0"
description = "Framework/toolkit for building Telegram bots with telebot and redis"
authors = ["Igor Vaiman <gosha.vaiman@gmail.com>"]
license = "GPLv3"
readme = "README.md"
repository = "https://github.com/bots-against-war/telebot-components"
packages = [{include = "telebot_components"}]

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
dirty = true
style="semver"

[tool.poetry.dependencies]
python = "^3.9"
telebot-against-war = "0.3.4"
redis = "^4.3.1"
py-trello = "^0.18.0"
markdownify = "^0.11.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-aiohttp = "1.0.4"
pytest-asyncio = "^0.18.3"
python-dotenv = "^0.20.0"
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
mypy = "^0.960"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
profile = 'black'

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    'integration: requires a real bot to run',
]
